Revert "do a bit of refactoring"
This reverts commit e4f743abdcde293eb1be891e4001b3467e79e67a.
